Annotations needed to create a skill moved out to a seperate project  to make an initial version of Skillup with easy use.
