+++ SkillUp-Annotations/src/main/java/skillup/annotations/Module.java
@@ -1,26 +1,19 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import actionGenerator.SkillType;
-
 /**
- * Skill Annotation needs skillIRI, moduleIRI and skill type which is an class
- * that extends from class SkillType. CapabilityIRI and description are
+ * Annotation Module needs a moduleIRI. Description and capabilityIRI are
  * optional.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface Skill {
-
-	public Class<? extends SkillType> type();
+public @interface Module {
 
 	public String description() default "";
-
-	public String skillIri();
 
 	public String moduleIri();
