+++ SkillUp-Annotations/src/main/java/skillup/annotations/OpcUaSkillType.java
@@ -1,4 +1,4 @@
-package actionGenerator;
+package skillup.annotations;
 
 /**
  * Class to tell that new skill is of type OPC-UA when this class is set as type
