+++ SkillUp-Annotations/src/main/java/skillup/annotations/Resetting.java
@@ -1,4 +1,4 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,6 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Stopping {
+public @interface Resetting {
 
 }
