+++ SkillUp-Annotations/src/main/java/skillup/annotations/Skill.java
@@ -1,4 +1,4 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,14 +6,19 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Annotation Module needs a moduleIRI. Description and capabilityIRI are
+ * Skill Annotation needs skillIRI, moduleIRI and skill type which is an class
+ * that extends from class SkillType. CapabilityIRI and description are
  * optional.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface Module {
+public @interface Skill {
+
+	public Class<? extends SkillType> type();
 
 	public String description() default "";
+
+	public String skillIri();
 
 	public String moduleIri();
