+++ SkillUp-Annotations/src/main/java/skillup/annotations/SkillParameter.java
@@ -1,4 +1,4 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,17 +6,19 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * SkillOutput annotation can have a name and description. Necessary is a
- * boolean value for isRequired which tells if this parameter is essential for
- * executing a skill method
+ * SkillParameter annotation can have a name, description and option. Necessary
+ * is a boolean value for isRequired which tells if this parameter is essential
+ * for executing a skill method
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
-public @interface SkillOutput {
+public @interface SkillParameter {
 
 	public String name() default "";
 
 	public String description() default "";
 
 	public boolean isRequired();
+
+	public String[] option() default "";
 }
