+++ SkillUp-Annotations/src/main/java/skillup/annotations/StateMachine.java
@@ -1,4 +1,4 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface Holding {
+@Target(ElementType.FIELD)
+public @interface StateMachine {
 
 }
