+++ SkillUp-Annotations/src/main/java/skillup/annotations/Unholding.java
@@ -1,4 +1,4 @@
-package annotations;
+package skillup.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
